Codechange: Move usage of ResolverObjects to newgrf_xxx.cpp
