+++ src/newgrf_roadstop.cpp
@@ -21,6 +21,7 @@
 #include "window_type.h"
 #include "timer/timer_game_calendar.h"
 #include "town.h"
+#include "tile_cmd.h"
 #include "viewport_func.h"
 #include "newgrf_animation_base.h"
 #include "newgrf_sound.h"
@@ -347,6 +348,14 @@ void DrawRoadStopTile(int x, int y, RoadType roadtype, const RoadStopSpec *spec,
 	}
 
 	DrawCommonTileSeqInGUI(x, y, dts, 0, 0, palette, true);
+}
+
+const TileLayoutSpriteGroup *GetRoadStopLayout(TileInfo *ti, const RoadStopSpec *spec, BaseStation *st, StationType type, int view)
+{
+	RoadStopResolverObject object(spec, st, ti->tile, INVALID_ROADTYPE, type, view);
+	const SpriteGroup *group = object.Resolve();
+	if (group == nullptr || group->type != SGT_TILELAYOUT) return nullptr;
+	return static_cast<const TileLayoutSpriteGroup *>(group);
 }
 
 /** Wrapper for animation control, see GetRoadStopCallback. */
