+++ src/station_cmd.cpp
@@ -3328,16 +3328,15 @@ draw_default_foundation:
 		if (stopspec != nullptr) {
 			stop_draw_mode = stopspec->draw_mode;
 			st = BaseStation::GetByTile(ti->tile);
-			RoadStopResolverObject object(stopspec, st, ti->tile, INVALID_ROADTYPE, type, view);
-			const SpriteGroup *group = object.Resolve();
-			if (group != nullptr && group->type == SGT_TILELAYOUT) {
+			const TileLayoutSpriteGroup *group = GetRoadStopLayout(ti, stopspec, st, type, view);
+			if (group != nullptr) {
 				if (stopspec->flags.Test(RoadStopSpecFlag::DrawModeRegister)) {
 					stop_draw_mode = static_cast<RoadStopDrawMode>(GetRegister(0x100));
 				}
 				if (type == StationType::RoadWaypoint && stop_draw_mode.Test(RoadStopDrawMode::WaypGround)) {
 					draw_ground = true;
 				}
-				t = ((const TileLayoutSpriteGroup *)group)->ProcessRegisters(nullptr);
+				t = group->ProcessRegisters(nullptr);
 			}
 		}
